@@ -4,16 +4,17 @@ gem 'active_model_serializers'
 gem 'devise_token_auth'
 gem 'omniauth'
 gem 'pg', '~> 0.15'
-gem 'rails', '4.2.5.1'
-gem 'rails-api'
+gem 'rails', '5.1.4'
 gem 'sdoc', '~> 0.4.0', group: :doc
 
 group :development, :test do
+  gem 'factory_bot_rails'
   gem 'pry'
   gem 'rspec-rails'
 end
 
 group :development do
+  gem 'listen'
   gem 'web-console', '~> 2.0'
   gem 'spring'
 end
