@@ -2,10 +2,12 @@ module Api
   module V1
     class SheetsController < ApplicationController
       before_action :set_sheet, :only => [:destroy, :show, :update]
-      before_action :authenticate_user!, :only => [:create, :update, :destroy]
-      before_action :permit_user_update, :only => [:update, :destroy]
+      # before_action :authenticate_user!, :only => [:create, :update, :destroy]
+      # before_action :permit_user_update, :only => [:update, :destroy]
 
       def create
+        # TODO: when devise is set up this won't be needed.  It sets current_user
+        current_user = User.find(sheet_params.fetch(:user_id))
         @sheet = current_user.sheets.build(sheet_params)
 
         if @sheet.save
@@ -34,7 +36,7 @@ module Api
 
       def update
         if @sheet.update_attributes(sheet_params)
-          render json: {}, :status => :no_content, :adapter => :json
+          render json: @sheet, :adapter => :json
         else
           render json: @sheet.errors, :status => :unprocessable_entity, :adapter => :json
         end
@@ -51,7 +53,7 @@ module Api
       end
 
       def sheet_params
-        params.require(:sheet).permit(:home_team, :away_team, :box_amount, :password, :name, :closed)
+        params.require(:sheet).permit(:home_team, :away_team, :box_amount, :password, :name, :closed, :game_id, :user_id)
       end
     end
   end
