@@ -1,13 +1,15 @@
-class Sheet < ActiveRecord::Base
+class Sheet < ApplicationRecord
   has_many :boxes, :dependent => :destroy
   belongs_to :user
   belongs_to :game
 
   validates :home_team, :away_team, :presence => true
 
-  before_create :init_boxes
-  before_create :generate_code
-  before_save :set_team_score_arrays, :if => Proc.new { |sheet| sheet.closed? && sheet.team_score_arrays_blank? }
+  def self.start_new_sheet(sheet_params)
+    instance = new(sheet_params.merge!(:sheet_code => self.generate_code))
+    self.init_boxes(instance)
+    return instance
+  end
 
   def belongs_to_user?(user)
     user_id == user.id
@@ -17,10 +19,6 @@ class Sheet < ActiveRecord::Base
     self.home_team_score_row = random_score_array
     self.away_team_score_row = random_score_array
     self.save
-  end
-
-  def random_score_array
-    (0..9).to_a.shuffle
   end
 
   def all_boxes_full?
@@ -33,16 +31,20 @@ class Sheet < ActiveRecord::Base
 
   private
 
-  def generate_code
-    self.sheet_code = ("A".."Z").to_a.shuffle.sample(4).join
+  def self.generate_code
+    ("A".."Z").to_a.shuffle.sample(4).join
   end
 
-  def init_boxes
+  def self.init_boxes(instance)
     # Creates 100 boxes with coordinates being a letter between A and J
     ids = ("A".."J").to_a
     id_array = ids.product(ids)
     id_array.each do |x_coord, y_coord|
-      self.boxes.build(:home_team_id => x_coord, :away_team_id => y_coord)
+      instance.boxes.build(:home_team_id => x_coord, :away_team_id => y_coord)
     end
+  end
+
+  def random_score_array
+    (0..9).to_a.shuffle
   end
 end
