@@ -1,4 +1,4 @@
-class CreateBoxes < ActiveRecord::Migration
+class CreateBoxes < ActiveRecord::Migration[5.1]
   def change
     create_table :boxes do |t|
       t.integer :home_team_num
