@@ -1,4 +1,4 @@
-class AddSheetRefToBoxes < ActiveRecord::Migration
+class AddSheetRefToBoxes < ActiveRecord::Migration[5.1]
   def change
     add_reference :boxes, :sheet, index: true, foreign_key: true
   end
