@@ -1,4 +1,4 @@
-class AddBoxAmountToSheets < ActiveRecord::Migration
+class AddBoxAmountToSheets < ActiveRecord::Migration[5.1]
   def change
     add_column :sheets, :box_amount, :integer
   end
