@@ -1,4 +1,4 @@
-class AddClosedToSheets < ActiveRecord::Migration
+class AddClosedToSheets < ActiveRecord::Migration[5.1]
   def change
     add_column :sheets, :closed, :boolean
   end
