@@ -1,4 +1,4 @@
-class AddUserToSheets < ActiveRecord::Migration
+class AddUserToSheets < ActiveRecord::Migration[5.1]
   def change
     add_reference :sheets, :user, index: true, foreign_key: true
   end
