@@ -1,4 +1,4 @@
-class CreateGames < ActiveRecord::Migration
+class CreateGames < ActiveRecord::Migration[5.1]
   def change
     create_table :games do |t|
       t.text :home_team
