@@ -7,5 +7,5 @@ Rails.application.routes.draw do
         resources :boxes, :only => [:update]
       end
     end
-  end
+  end  # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
 end
